Further DHT troubleshooting: general validator/worker/user-queries and refactoring
